Big refactor, moved models and views to separate files.
