+++ src/js/main.js
@@ -1,130 +1,23 @@
 var $ = require('jquery');
-var fs = require('fs');
 var _ = require('underscore');
 var Backbone = require('backbone');
+var fs = require('fs');
 var handlebars = require('handlebars');
 
 // Backbone assumes jQuery is in the global namespace
 Backbone.$ = $;
 
+// TODO: Stop hardcoding host, configure dev/staging/prod host with gulp
 $.ajaxPrefilter(function( options, originalOptions, jqXHR) {
   options.url = "http://localhost:8080/api" + options.url;
 });
 
+// Models
+var User = require('./models/user.js');
 
-var User = Backbone.Model.extend({
-  defaults: {
-    username: '',
-    id: ''
-  },
-
-  url: '/user'
-});
-
-var Search = Backbone.Model.extend({
-  defaults: {
-    data: {}
-  },
-
-  url: '/search'
-});
-
-var SearchFormView = Backbone.View.extend({
-  initialize: function() {
-    this.render();
-  },
-
-  el: $('#search-form-container'),
-
-  template: function() {
-    var templateFile = fs.readFileSync('src/templates/search-form.hbs', 'utf8');
-    return handlebars.compile(templateFile);
-  },
-
-  render: function(data) {
-    var html = this.template()
-    this.$el.append(html);
-  },
-
-  events: {
-    "submit": "search",
-  },
-
-  search: function(e) {
-    e.preventDefault();
-    var q = 'q=' + $('#query').val(),
-        lang = 'lang='  + $('#lang').val(),
-        result_type = 'result_type=' + $('#result-type').val();
-
-    var search = new Search();
-    search.fetch({
-      data: [q, lang, result_type].join('&'),
-      success: function(data) {
-        var resultsTableView = new ResultsTableView();
-        resultsTableView.render(data.toJSON());
-      }
-    });
-  }
-});
-
-// View for rendering a list of tweets
-var ResultsTableView = Backbone.View.extend({
-  el: $('#results-container'),
-
-  initialize: function() {
-    var self = this;
-    this.render();
-  },
-
-  render: function (data) {
-    var tweets = [];
-    _.each(data, function(k,v) {
-      tweets.push(k);
-    });
-    var data = {
-      statuses: tweets
-    };
-
-    // Dummy data
-    // var mockData = fs.readFileSync('src/js/search-response.json', 'utf8'),
-    //     response = JSON.parse(mockData);
-
-    var templateFile = fs.readFileSync('src/templates/tweet.hbs', 'utf8'),
-        template = handlebars.compile(templateFile);
-
-    var html = template(data);
-    console.log(html);
-    this.$el.html(html);
-    return this;
-  }
-});
-
-var LoginFormView = Backbone.View.extend({
-  initialize: function() {
-    this.render();
-  },
-  el: $('#login-form-container'),
-
-  events: {
-    "submit": "login"
-  },
-
-  template: function() {
-    var templateFile = fs.readFileSync('src/templates/login-form.hbs', 'utf8');
-    return handlebars.compile(templateFile);
-  },
-
-  login: function(e) {
-    e.preventDefault();
-    $(location).attr('href','http://localhost:8080/api/oauth');
-  },
-
-  render: function() {
-    var html = this.template();
-    this.$el.append(html);
-    return this;
-  },
-});
+// Views
+var SearchFormView = require('./views/search-form.js');
+var LoginFormView = require('./views/login-form.js');
 
 var Router = Backbone.Router.extend({
   routes: {
@@ -139,8 +32,6 @@ var Router = Backbone.Router.extend({
     var self = this;
 
     this.on('route:home', function() {
-      // this.searchFormView = new SearchFormView();
-      // this.resultsTableView = new ResultsTableView();
       this.currentUser = new User();
       this.currentUser.fetch({
         success: function(user) {
@@ -156,10 +47,5 @@ var Router = Backbone.Router.extend({
 });
 
 var router = new Router();
-
-// var resultsTableView = new ResultsTableView();
-// var headerView = new SearchFormView();
-
-var app = {}
 
 Backbone.history.start();
