+++ src/js/models/search.js
@@ -1,0 +1,12 @@
+var $ = require('jquery');
+var _ = require('underscore');
+var Backbone = require('backbone');
+
+module.exports = Backbone.Model.extend({
+  defaults: {
+    data: {}
+  },
+
+  url: '/search'
+});
+
