+++ src/js/models/user.js
@@ -1,0 +1,13 @@
+var $ = require('jquery');
+var _ = require('underscore');
+var Backbone = require('backbone');
+
+module.exports = Backbone.Model.extend({
+  defaults: {
+    username: '',
+    id: ''
+  },
+
+  url: '/user'
+});
+
